+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * Add char-at, which wraps the native String/charAt
+* Add re-quote which quotes a string, for use in regular expressions.
 
 ### Bugs fixed